requirements.txt
Flask
bcrypt
pyjwt
werkzeug